"Progressively" enhance navigating to `seats#show`
For comparison [see this commit][commit]

Note that "Progressively" is in quotes, because while this represents a small
iterative change that uses UJS, its not truly progressive enhancement in the
Rails sense.

The referenced commit attempts to efficiently render a tiny portion
(modal contents) of a page without rendering everything else. To quote:

> During a full page visit to the `seats#show` action via a direct `GET
> /venues/:venue_id/floors/:floor_id/seats/:id` request, that action
> fetches the entire set of associated `Seat` records for a given `Floor`,
> and renders the entire set into the template.
>
> In this application's case, that's between 1,300 and 1,500 records that
> are fetched, and as many (or more) corresponding [`<svg>`
> elements][mdn-svg] sent to the browser.

Here we begin to highlight Superglue's more unique features. Your `show.json.props`
is a queryable content tree powered by [props_template]. It is possible to pass in
a path to a target node, traverse the template tree you built, while skipping
execution of nodes that do not matter.

To do this we append a `props_at` parameter in a link enabled with `data-sg-visit`
like so:

```
  <a href='/venues/benedum_center/floors/orchestra/seats/B-103?props_at=data.seat'
    data-sg-visit={true}
    data-sg-placeholder={this.props.pageKey}
  />
```

or using the equivalent provided thunk like so:

```
  this.props.visit('/venues/benedum_center/floors/orchestra/seats/B-103?props_at=data.seat', {
    placeholderKey: this.props.pageKey
  })
```

In both scenarios, a placeholderKey is passed to the `visit` thunk, allowing
you to copy a specified page as a placeholder for the next page. Then
Superglue fetches the node from the server, and finally grafts it onto the copy at
the same location.

This is functionality that is built into Superglue. Turbolinks / Stimulus solutions
may be able to achieve this in a variety of ways, but would require additional
work. Comparing the two approaches, Superglue's out-of-the-box support for
[same-page] updates just has better prose.

Lets take a look at how we can achieve this in Stimulus/ Turbolinks:

1. The referenced commit uses a custom header and a variant. Variants are
usually used to show different versions of the page based on device type. In
this commits case a variant along with a custom header to decide to render a
modal or a full page load.

2. Make an API, and do the modal client side. This would require a brand new set
of routes, views, tests, and more work than necessary.

[commit]: https://github.com/seanpdoyle/select-your-own-seat/commit/1c1591fda28eeccdba35ec64db75a1c4d272c0bb
[same-page]: https://thoughtbot.github.io/superglue/#/recipes/load-tab-onclick
[props_template]: https://github.com/thoughtbot/props_template