Refactor: Extract SeatingMap component [NEW]
This commit has no equivalent in the Stimulus branch. Here, a concept of a
`SeatingMap` component emerges from `index.js`. This is a quality of working in
React. The library library makes it easy to extract related and semantic behavior
into into reusable, testable components.

Rails has tools like partial, presenters, view object, etc. But when faced with
a ERB template with a rich history of changes, dependencies on instance variables,
global view helpers, local variables contribute to some of the difficulty of
extracting isolated, and testable behavior.