Manage filter query parameters with push state
For comparison [see this commit][commit]

There is a slight difference here on how to get this functionality to work. In
the Stimulus version, its a change to the `window.history` when a specific filter
is selected. In the Superglue version, we have to copy the current state to the
next url, modify the maximum on that copy, then call a utility function
that gets passed to your component, `navigateTo`.

While the Superglue version is a bit lengthier, it makes "undoing" this action
easy, as `history.back` would properly revert that state.

[commit]: https://github.com/seanpdoyle/select-your-own-seat/commit/6236e5b81d0d4de460375af2cf8e24b3ea70090e