Add `seats#show` route and action
For comparison [see this commit][commit]

This changes here has a lot in common with the referenced commit. The usual
routes, controllers, models, and tests remain unchanged. There are some
differences in the views worth mentioning.

1. `SeatDialog.js` vs. `_dialog.html.erb` - React components vs. partials +
Stimulus. They're both using the same dialogPolyfill, ergonomically they're not
too dissimilar. The stimulus version uses the DOM api, the React version uses
its lifecycle.

A quality of the Stimulus version is it immediately starts out as HTML, and
behavior is added as sprinkles of javascript.

A quality of the React version is its testability, and its tendency to encourage
grouping of behavior into reusable components.

Its important to note that in both version, there are no additional HTTP request
beyond the initial load.

2. There is no `show.js`
In Superglue applications, there is usually a `js` companion to the `props`
template, but in this case it was left out intentionally. Instead, we reuse
the `index.js` because the only difference between the `index` and the `show`
is the modal. Here's the relevant line:

```
export const pageIdentifierToPageComponent = {
-  'seats/index': SeatsIndex
+  'seats/index': SeatsIndex,
+  'seats/show': SeatsIndex,
};
```

[commit]: https://github.com/seanpdoyle/select-your-own-seat/commit/438868d52abda9ef4d343a931878c0bef918a3d6