Switch floors from map
For comparison [see this commit][commit]

The commit above adds `active_link_to`. A gem that adds classes to links if
they share the same path with the current_url.

We can use the gem too! This commit uses a different helper from the same
library to achieve the same effect.

Co-authored-by: Jacob Galati <jacob@realfoto.ca>

[commit]: https://github.com/seanpdoyle/select-your-own-seat/commit/1f53ca717a6c3ab1f4e7219098f9c74101b1daf8