Dynamically Hide and Show Loading Spinner
For comparison [see this commit][commit]

This commit is a slight divergence from Stimulus. Superglue does not have global
callbacks like `turbolinks:request-start`, if you need to do something before a
request starts, you must do so explicitly.

The behavior of showing a loading indicator requires an optimistic update. We
have to change some loading state to `true` before we make a request, and
change it to `false` when we're done with the request.

To make this feature work, we're going to use Redux slices, an escape for when
we need frontend state management. The Superglue installer already sets up a
place where you can put this logic.

Much like the `flash` slice, we'll create a redux slice (call it `loading`) and
add customizations to the generated `application_visit.js`.

```
  reducers: {
    hideLoading() {
      return {
        isLoading: false
      }
    },
    showLoading() {
      return {
        isLoading: true
      }
    },
  },
```

[commit]: https://github.com/seanpdoyle/select-your-own-seat/commit/71697874c5109f55405a9d2ab29a538ed1f12c71