"Progressively" enhance price filtering
For comparison [see this commit][commit]

The goals of this commit is exactly the same as the referenced commit.
Implementation is different, with Superglue we have local state via react that
overrides the server sent maximum, and later, a function `filterSectionsByMax`
handles the selection by the state's new maximum.

There is a difference here that is worthy of note:

The `filterSectionsByMax` is a port of the server's logic located in
`_seating_map.json.props`. For reference:

```ruby
    if params.fetch(:maximum, Float::INFINITY).to_f >= section.price
      json.opacity "1.0"
      json.hidden false
    else
      json.opacity "0.3"
      json.hidden true
    end
```

and the JS version

```javascript
const filterSectionsByMax = (sections, maximum) => {
  return sections.map((section) => {
    const {opacity, hidden} = section
    let nextOpacity, nextHidden;

    if((maximum || Infinity)>= section.price) {
      nextOpacity = '1.0'
      nextHidden = false
    } else {
      nextOpacity = '0.3'
      nextHidden = true
    }

   if (opacity === nextOpacity && hidden === nextHidden) {
      return section
    } else  {
      return {...section, hidden: nextHidden, opacity: nextOpacity}
  }
}
```

[commit]: [https://github.com/seanpdoyle/select-your-own-seat/commit/d28a1a560df65677015cae331d2f375b22e530b3